src/03-rag-multi-vector-pyd/requirements.txt
openai==1.42.0
langchain-core==0.2.29
pydantic==2.8.2
python-dotenv==1.0.1
tiktoken==0.7.0
azure-search-documents==11.5.1
azure-identity==1.17.1
streamlit==1.37.1
promptflow==1.15.0
promptflow-core==1.15.0
langchain-community==0.2.12
langchain-openai==0.1.20
opentelemetry-instrumentation-langchain==0.26.5
langchain==0.2.14
langchain-text-splitters==0.2.2
langchain-qdrant==0.1.3